Refactor DataProviderSynchronizer: extract per-event persistence handlers (#804)
The synchronizer mixed load-bearing queue plumbing (subscription, startup
reclaim, drain serialization, dead-letter surfacing, retry/backoff) with a
~70-line manual switch and 15 per-event EF persistence methods, so adding a
persisted event meant editing the switch and the class, and an unmapped type
failed at runtime.

Extract the per-event persistence into discrete IPlayerUpdateHandler<TEvent>
classes (one per event), resolved per drain scope by PlayerUpdateEventDispatcher
keyed on the envelope's type name. Adding a persisted event is now a new handler
plus one AddPlayerUpdateHandler registration — the consume-side mirror of the
publish-side PlayerPersistencePublisher registration — rather than a switch edit.
The synchronizer keeps only the queue/drain/retry/dead-letter coordination and
delegates each apply to the dispatcher.

Behaviour is unchanged: every persistence body and comment is preserved verbatim,
the unknown-type and malformed-payload poison signals still dead-letter without
retrying, and the full DataProviderSynchronizer integration suite passes. Adds a
focused unit test for the dispatcher in isolation (now independently testable),
and documents the symmetric consume-side dispatch in backend-persistence.md.
