+++ Game.DataAccess/DataProviderSynchronizer.cs
@@ -1,9 +1,6 @@
-using Game.Infrastructure.Entities;
 using Game.Abstractions.Infrastructure;
 using Game.Core;
-using Game.Core.Players.Events;
-using Game.Infrastructure.Database;
-using Microsoft.EntityFrameworkCore;
+using Game.DataAccess.PlayerUpdates;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -213,379 +210,15 @@ namespace Game.DataAccess
             }
         }
 
+        // Applies a single dequeued event to the database. A fresh scope per call (and per retry attempt)
+        // gives each apply its own GameContext, and the dispatcher resolves the discrete
+        // IPlayerUpdateHandler for the envelope's type — an unregistered type surfaces as an
+        // UnknownEventTypeException (dead-lettered without retry by ProcessMessage above).
         private async Task HandleEvent(DomainEventEnvelope envelope)
         {
             using var scope = _services.CreateScope();
-            var context = scope.ServiceProvider.GetRequiredService<GameContext>();
-
-            switch (envelope.Type)
-            {
-                case nameof(PlayerCoreUpdatedEvent):
-                    var coreEvt = Deserialize<PlayerCoreUpdatedEvent>(envelope.Payload);
-                    await HandlePlayerCoreUpdated(context, coreEvt);
-                    break;
-
-                case nameof(AttributeAllocationsChangedEvent):
-                    var attrEvt = Deserialize<AttributeAllocationsChangedEvent>(envelope.Payload);
-                    await HandleAttributeAllocationsChanged(context, attrEvt);
-                    break;
-
-                case nameof(ItemUnlockedEvent):
-                    var unlockEvt = Deserialize<ItemUnlockedEvent>(envelope.Payload);
-                    await HandleItemUnlocked(context, unlockEvt);
-                    break;
-
-                case nameof(ItemEquippedEvent):
-                    var equipEvt = Deserialize<ItemEquippedEvent>(envelope.Payload);
-                    await HandleItemEquipped(context, equipEvt);
-                    break;
-
-                case nameof(ItemUnequippedEvent):
-                    var unequipEvt = Deserialize<ItemUnequippedEvent>(envelope.Payload);
-                    await HandleItemUnequipped(context, unequipEvt);
-                    break;
-
-                case nameof(ModUnlockedEvent):
-                    var modUnlockEvt = Deserialize<ModUnlockedEvent>(envelope.Payload);
-                    await HandleModUnlocked(context, modUnlockEvt);
-                    break;
-
-                case nameof(ModAppliedEvent):
-                    var modApplyEvt = Deserialize<ModAppliedEvent>(envelope.Payload);
-                    await HandleModApplied(context, modApplyEvt);
-                    break;
-
-                case nameof(ModRemovedEvent):
-                    var modRemoveEvt = Deserialize<ModRemovedEvent>(envelope.Payload);
-                    await HandleModRemoved(context, modRemoveEvt);
-                    break;
-
-                case nameof(SkillUnlockedEvent):
-                    var skillUnlockEvt = Deserialize<SkillUnlockedEvent>(envelope.Payload);
-                    await HandleSkillUnlocked(context, skillUnlockEvt);
-                    break;
-
-                case nameof(SelectedSkillsChangedEvent):
-                    var selectedSkillsEvt = Deserialize<SelectedSkillsChangedEvent>(envelope.Payload);
-                    await HandleSelectedSkillsChanged(context, selectedSkillsEvt);
-                    break;
-
-                case nameof(ItemFavoriteChangedEvent):
-                    var favoriteEvt = Deserialize<ItemFavoriteChangedEvent>(envelope.Payload);
-                    await HandleItemFavoriteChanged(context, favoriteEvt);
-                    break;
-
-                case nameof(LogPreferenceChangedEvent):
-                    var logEvt = Deserialize<LogPreferenceChangedEvent>(envelope.Payload);
-                    await HandleLogPreferenceChanged(context, logEvt);
-                    break;
-
-                case nameof(ProgressUpdatedEvent):
-                    var progressEvt = Deserialize<ProgressUpdatedEvent>(envelope.Payload);
-                    await HandleProgressUpdated(context, progressEvt);
-                    break;
-
-                // PlayerLeveledUpEvent is handled in-process only — it has no persistence
-                // handler registered, so it is never published to this queue.
-
-                default:
-                    throw new UnknownEventTypeException(envelope.Type);
-            }
+            var dispatcher = new PlayerUpdateEventDispatcher(scope.ServiceProvider);
+            await dispatcher.DispatchAsync(envelope);
         }
-
-        private static async Task HandlePlayerCoreUpdated(GameContext context, PlayerCoreUpdatedEvent evt)
-        {
-            await context.Players
-                .Where(p => p.Id == evt.PlayerId)
-                .ExecuteUpdateAsync(s => s
-                    .SetProperty(p => p.Level, evt.Level)
-                    .SetProperty(p => p.Exp, evt.Exp)
-                    .SetProperty(p => p.CurrentZoneId, evt.CurrentZoneId)
-                    .SetProperty(p => p.StatPointsGained, evt.StatPointsGained)
-                    .SetProperty(p => p.StatPointsUsed, evt.StatPointsUsed));
-        }
-
-        private static async Task HandleAttributeAllocationsChanged(GameContext context, AttributeAllocationsChangedEvent evt)
-        {
-            var currentRows = await context.PlayerAttributes
-                .Where(pa => pa.PlayerId == evt.PlayerId)
-                .ToListAsync();
-
-            var rowsByAttributeId = currentRows.ToDictionary(pa => pa.AttributeId);
-
-            foreach (var alloc in evt.Allocations)
-            {
-                var attributeId = (int)alloc.Attribute;
-                var amount = (decimal)alloc.Amount;
-
-                if (rowsByAttributeId.TryGetValue(attributeId, out var row))
-                {
-                    if (amount == 0)
-                    {
-                        context.PlayerAttributes.Remove(row);
-                    }
-                    else
-                    {
-                        row.Amount = amount;
-                    }
-                }
-                else if (amount != 0)
-                {
-                    context.PlayerAttributes.Add(new PlayerAttribute
-                    {
-                        PlayerId = evt.PlayerId,
-                        AttributeId = attributeId,
-                        Amount = amount,
-                    });
-                }
-            }
-
-            await context.SaveChangesAsync();
-        }
-
-        private static async Task HandleItemUnlocked(GameContext context, ItemUnlockedEvent evt)
-        {
-            var exists = await context.UnlockedItems
-                .AnyAsync(ui => ui.PlayerId == evt.PlayerId && ui.ItemId == evt.ItemId);
-
-            if (!exists)
-            {
-                context.UnlockedItems.Add(new UnlockedItem
-                {
-                    PlayerId = evt.PlayerId,
-                    ItemId = evt.ItemId,
-                });
-                await context.SaveChangesAsync();
-            }
-        }
-
-        private static async Task HandleItemEquipped(GameContext context, ItemEquippedEvent evt)
-        {
-            // Absolute per-slot upsert in a single atomic UPDATE: the equipped item takes the slot and any
-            // prior occupant is cleared, in one statement. The previous two-statement clear-then-set left a
-            // window where a crash between the writes (the queue read is a destructive pop, so the event is
-            // not redelivered) emptied the slot for good. Folding both into one CASE-driven update over the
-            // affected rows removes that window and stays idempotent — re-applying converges to the same state,
-            // including when the item is moving from another slot (its own row is reassigned, vacating the old).
-            await context.UnlockedItems
-                .Where(ui => ui.PlayerId == evt.PlayerId && (ui.ItemId == evt.ItemId || ui.EquipmentSlotId == evt.SlotId))
-                .ExecuteUpdateAsync(s => s.SetProperty(
-                    ui => ui.EquipmentSlotId,
-                    ui => ui.ItemId == evt.ItemId ? evt.SlotId : (int?)null));
-        }
-
-        private static async Task HandleItemUnequipped(GameContext context, ItemUnequippedEvent evt)
-        {
-            await context.UnlockedItems
-                .Where(ui => ui.PlayerId == evt.PlayerId && ui.ItemId == evt.ItemId)
-                .ExecuteUpdateAsync(s => s.SetProperty(ui => ui.EquipmentSlotId, (int?)null));
-        }
-
-        private static async Task HandleModUnlocked(GameContext context, ModUnlockedEvent evt)
-        {
-            var exists = await context.UnlockedMods
-                .AnyAsync(um => um.PlayerId == evt.PlayerId && um.ItemModId == evt.ItemModId);
-
-            if (!exists)
-            {
-                context.UnlockedMods.Add(new UnlockedMod
-                {
-                    PlayerId = evt.PlayerId,
-                    ItemModId = evt.ItemModId,
-                });
-                await context.SaveChangesAsync();
-            }
-        }
-
-        private static async Task HandleSkillUnlocked(GameContext context, SkillUnlockedEvent evt)
-        {
-            var exists = await context.PlayerSkills
-                .AnyAsync(ps => ps.PlayerId == evt.PlayerId && ps.SkillId == evt.SkillId);
-
-            if (!exists)
-            {
-                // Earning a skill unlocks it without equipping it: Selected = false, Order = 0
-                // (the player chooses their loadout separately). Idempotent insert mirrors the
-                // item/mod unlock handlers so re-applying the event never duplicates the row.
-                context.PlayerSkills.Add(new PlayerSkill
-                {
-                    PlayerId = evt.PlayerId,
-                    SkillId = evt.SkillId,
-                    Selected = false,
-                    Order = 0,
-                });
-                await context.SaveChangesAsync();
-            }
-        }
-
-        private static async Task HandleSelectedSkillsChanged(GameContext context, SelectedSkillsChangedEvent evt)
-        {
-            // Delete-then-rebuild for idempotency, applied as a single write (the same shape as the
-            // attribute-allocations handler): fetch the player's skill rows, reset every flag, then mark
-            // each id in the ordered loadout Selected = true with its index as Order. Re-applying the event
-            // converges to the same state, and EF batches the touched rows into one round-trip rather than
-            // issuing one ExecuteUpdate per skill.
-            var playerSkills = await context.PlayerSkills
-                .Where(ps => ps.PlayerId == evt.PlayerId)
-                .ToListAsync();
-
-            var orderBySkillId = new Dictionary<int, int>(evt.OrderedSkillIds.Count);
-            for (var index = 0; index < evt.OrderedSkillIds.Count; index++)
-            {
-                orderBySkillId[evt.OrderedSkillIds[index]] = index;
-            }
-
-            foreach (var playerSkill in playerSkills)
-            {
-                if (orderBySkillId.TryGetValue(playerSkill.SkillId, out var order))
-                {
-                    playerSkill.Selected = true;
-                    playerSkill.Order = order;
-                }
-                else
-                {
-                    playerSkill.Selected = false;
-                    playerSkill.Order = 0;
-                }
-            }
-
-            await context.SaveChangesAsync();
-        }
-
-        private static async Task HandleModApplied(GameContext context, ModAppliedEvent evt)
-        {
-            // Remove existing mod in the same slot
-            await context.AppliedMods
-                .Where(am => am.PlayerId == evt.PlayerId && am.ItemId == evt.ItemId && am.ItemModSlotId == evt.ItemModSlotId)
-                .ExecuteDeleteAsync();
-
-            context.AppliedMods.Add(new AppliedMod
-            {
-                PlayerId = evt.PlayerId,
-                ItemId = evt.ItemId,
-                ItemModSlotId = evt.ItemModSlotId,
-                ItemModId = evt.ItemModId,
-            });
-            await context.SaveChangesAsync();
-        }
-
-        private static async Task HandleModRemoved(GameContext context, ModRemovedEvent evt)
-        {
-            await context.AppliedMods
-                .Where(am => am.PlayerId == evt.PlayerId && am.ItemId == evt.ItemId && am.ItemModSlotId == evt.ItemModSlotId)
-                .ExecuteDeleteAsync();
-        }
-
-        private static async Task HandleItemFavoriteChanged(GameContext context, ItemFavoriteChangedEvent evt)
-        {
-            // Idempotent absolute update: set the flag to the event's value for the player's unlocked item.
-            await context.UnlockedItems
-                .Where(ui => ui.PlayerId == evt.PlayerId && ui.ItemId == evt.ItemId)
-                .ExecuteUpdateAsync(s => s.SetProperty(ui => ui.Favorite, evt.Favorite));
-        }
-
-        private static async Task HandleLogPreferenceChanged(GameContext context, LogPreferenceChangedEvent evt)
-        {
-            var logTypeId = (int)evt.LogType;
-
-            // Idempotent upsert mirroring HandleItemFavoriteChanged: attempt the absolute update first as a
-            // single self-committing write; if no row exists yet (rows-affected 0) fall through to the insert.
-            // Re-applying the event converges to the same state under the write-behind retry policy.
-            var updated = await context.LogPreferences
-                .Where(lp => lp.PlayerId == evt.PlayerId && lp.LogTypeId == logTypeId)
-                .ExecuteUpdateAsync(s => s.SetProperty(lp => lp.Enabled, evt.Enabled));
-
-            if (updated == 0)
-            {
-                context.LogPreferences.Add(new LogPreference
-                {
-                    PlayerId = evt.PlayerId,
-                    LogTypeId = logTypeId,
-                    Enabled = evt.Enabled,
-                });
-                await context.SaveChangesAsync();
-            }
-        }
-
-        private static async Task HandleProgressUpdated(GameContext context, ProgressUpdatedEvent evt)
-        {
-            // Absolute upserts so re-applying the event under the retry policy converges to the same state.
-            // Batched like HandleAttributeAllocationsChanged: load the touched rows, set/insert, save once.
-            if (evt.Statistics.Count > 0)
-            {
-                // Bound the load by the touched type id set AND the touched entity id set — their
-                // cross-product, which is a superset of the exact (type, entity) pairs changed. Filtering on
-                // typeIds alone would, for a long-lived account with one row per enemy/skill, load hundreds to
-                // upsert the ~10-20 this battle changed (aggregate-DB-load concern, #548). A value-tuple IN
-                // over the exact pairs isn't cleanly EF-translatable, so this cross-product bound is the
-                // pragmatic narrowing; the exact-key match still happens in memory below. entityIds includes
-                // null for the global rows — EF turns Contains over a List<int?> into
-                // "EntityId IN (...) OR EntityId IS NULL".
-                var typeIds = evt.Statistics.Select(s => s.StatisticTypeId).Distinct().ToList();
-                var entityIds = evt.Statistics.Select(s => s.EntityId).Distinct().ToList();
-                var existing = await context.PlayerStatistics
-                    .Where(ps => ps.PlayerId == evt.PlayerId
-                        && typeIds.Contains(ps.StatisticTypeId)
-                        && entityIds.Contains(ps.EntityId))
-                    .ToListAsync();
-                var byKey = existing.ToDictionary(ps => (ps.StatisticTypeId, ps.EntityId));
-
-                foreach (var stat in evt.Statistics)
-                {
-                    if (byKey.TryGetValue((stat.StatisticTypeId, stat.EntityId), out var row))
-                    {
-                        row.Value = stat.Value;
-                    }
-                    else
-                    {
-                        context.PlayerStatistics.Add(new PlayerStatistic
-                        {
-                            PlayerId = evt.PlayerId,
-                            StatisticTypeId = stat.StatisticTypeId,
-                            EntityId = stat.EntityId,
-                            Value = stat.Value,
-                        });
-                    }
-                }
-            }
-
-            if (evt.Challenges.Count > 0)
-            {
-                var challengeIds = evt.Challenges.Select(c => c.ChallengeId).ToList();
-                var existing = await context.PlayerChallenges
-                    .Where(pc => pc.PlayerId == evt.PlayerId && challengeIds.Contains(pc.ChallengeId))
-                    .ToListAsync();
-                var byId = existing.ToDictionary(pc => pc.ChallengeId);
-
-                foreach (var challenge in evt.Challenges)
-                {
-                    if (byId.TryGetValue(challenge.ChallengeId, out var row))
-                    {
-                        row.Progress = challenge.Progress;
-                        row.Completed = challenge.Completed;
-                        row.CompletedAt = challenge.CompletedAt;
-                    }
-                    else
-                    {
-                        context.PlayerChallenges.Add(new PlayerChallenge
-                        {
-                            PlayerId = evt.PlayerId,
-                            ChallengeId = challenge.ChallengeId,
-                            Progress = challenge.Progress,
-                            Completed = challenge.Completed,
-                            CompletedAt = challenge.CompletedAt,
-                        });
-                    }
-                }
-            }
-
-            await context.SaveChangesAsync();
-        }
-
-        private static T Deserialize<T>(string json)
-            => json.Deserialize<T>() ?? throw new JsonException($"Deserialized '{typeof(T).Name}' payload was null.");
-
-        private sealed class UnknownEventTypeException(string eventType)
-            : Exception($"Unrecognized player event type '{eventType}'.");
     }
 }
