+++ docs/backend-persistence.md
@@ -14,7 +14,7 @@ The TTL is sized so eviction can't bite mid-drain: because the cache is the sour
 
 ## The write-behind path
 
-When the `Player` aggregate mutates it raises domain events; on save, the repository dispatches them in-process and a publisher serializes each into an envelope. Rather than publishing per event, the publisher **buffers the envelopes and flushes the whole batch onto the Redis queue as one multi-value push** after dispatch settles, so a save raising several events costs a single queue round-trip. A background service consumes that queue and applies each change to the database. Only events representing persistent state changes are registered for publishing; purely in-process reactions are never enqueued — keeping the "which events persist" decision in the data-access layer rather than encoded onto the domain events.
+When the `Player` aggregate mutates it raises domain events; on save, the repository dispatches them in-process and a publisher serializes each into an envelope. Rather than publishing per event, the publisher **buffers the envelopes and flushes the whole batch onto the Redis queue as one multi-value push** after dispatch settles, so a save raising several events costs a single queue round-trip. A background service consumes that queue and applies each change to the database. The consume side mirrors the publish side: each enqueued event type has a discrete persistence handler, resolved per drain by a dispatcher keyed on the envelope's type name, so adding a persisted event is a new handler plus one registration rather than an edit to a central switch (an envelope whose type matches no handler is dead-lettered as a poison message). Only events representing persistent state changes are registered for publishing; purely in-process reactions are never enqueued — keeping the "which events persist" decision in the data-access layer rather than encoded onto the domain events.
 
 **The source-of-truth cache write is deliberately fire-and-forget**, because the in-memory aggregate — not the cache — is the read-modify-write base during a session. A live player runs on one WebSocket whose request-scoped session lives for the whole connection; the aggregate is loaded from Redis **once** up front and every later command mutates and re-saves that same in-memory object rather than re-reading Redis. So a dropped cache write can't poison the next command (it never reads the key) and self-heals on the next save, while Postgres still converges via the queue. The one residual exposure — a dropped write *combined with* the in-memory aggregate being discarded (reconnect / socket-replace) before the next save — is a knowingly-accepted trade-off, since guarding it would put an awaited Redis round-trip on the per-battle hot path.
